Update dev dependencies
Switch from spaces to tabs

Bump patch version --> 0.4.1
