+++ .eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": [
-    "eslint-config-carnesen"
-  ]
+	"extends": [
+		"eslint-config-carnesen"
+	]
 }
