+++ .npmrc
@@ -1,2 +1,3 @@
 package-lock = true
+save-dev = true
 save-exact = true
