+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
-  "typescript.tsdk": "node_modules/typescript/lib",
-  "cSpell.words": [
-    "carnesen"
-  ]
+	"typescript.tsdk": "node_modules/typescript/lib",
+	"cSpell.words": [
+		"carnesen"
+	]
 }
