+++ changelog.md
@@ -1,9 +1,20 @@
-# `@carnesen/run-and-catch` change log
+# **@carnesen/run-and-catch** changelog
+
+## 0.4.1
+
+## Internal
+
+- Update dev dependencies
+- Use tabs instead of spaces
 
 ## 0.4.0
+
 ### Breaking
+
 - Drop support for Node.js version 8
+
 ### Internal
+
 - Update dev dependencies
 - Use GitHub Actions for CI
 - Use ESLint for linting
