+++ example.js
@@ -2,20 +2,22 @@
 const { runAndCatch } = require('.');
 
 function throwAnError(message) {
-  throw new Error(message);
+	throw new Error(message);
 }
 
-function doNothing() { }
+function doNothing() {}
 
 (async () => {
-  const returnValue = await runAndCatch(throwAnError, 'something bad happened');
-  console.log(`Q: Return value is an Error? A: ${returnValue instanceof Error}.`);
-  console.log(`returnValue.message: "${returnValue.message}"`);
-  try {
-    await runAndCatch(doNothing);
-  } catch (exception) {
-    console.log(`Exception message: ${exception.message}`);
-  }
+	const returnValue = await runAndCatch(throwAnError, 'something bad happened');
+	console.log(
+		`Q: Return value is an Error? A: ${returnValue instanceof Error}.`,
+	);
+	console.log(`returnValue.message: "${returnValue.message}"`);
+	try {
+		await runAndCatch(doNothing);
+	} catch (exception) {
+		console.log(`Exception message: ${exception.message}`);
+	}
 })();
 // Q: Return value is an Error? A: true.
 // returnValue.message: "something bad happened"
