+++ jest.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  preset: 'ts-jest',
+	preset: 'ts-jest',
 };
